Move common function to detect new datasets to utils cached
